--- a/usr/src/lib/libzfs/common/libzfs_dataset.c
+++ b/usr/src/lib/libzfs/common/libzfs_dataset.c
@@ -786,7 +786,8 @@ libzfs_mnttab_cache_compare(const void *arg1, const void *arg2)
 void
 libzfs_mnttab_init(libzfs_handle_t *hdl)
 {
-	(void) mutex_init(&hdl->libzfs_mnttab_cache_lock, USYNC_THREAD, NULL);
+	(void) mutex_init(&hdl->libzfs_mnttab_cache_lock,
+	    LOCK_NORMAL | LOCK_ERRORCHECK, NULL);
 	assert(avl_numnodes(&hdl->libzfs_mnttab_cache) == 0);
 	avl_create(&hdl->libzfs_mnttab_cache, libzfs_mnttab_cache_compare,
 	    sizeof (mnttab_node_t), offsetof(mnttab_node_t, mtn_node));
@@ -858,7 +859,7 @@ libzfs_mnttab_find(libzfs_handle_t *hdl, const char *fsname,
 			return (ENOENT);
 	}
 
-	(void) mutex_lock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_enter(&hdl->libzfs_mnttab_cache_lock);
 	if (avl_numnodes(&hdl->libzfs_mnttab_cache) == 0)
 		libzfs_mnttab_update(hdl);
 
@@ -868,7 +869,7 @@ libzfs_mnttab_find(libzfs_handle_t *hdl, const char *fsname,
 		*entry = mtn->mtn_mt;
 		ret = 0;
 	}
-	(void) mutex_unlock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_exit(&hdl->libzfs_mnttab_cache_lock);
 	return (ret);
 }
 
@@ -878,7 +879,7 @@ libzfs_mnttab_add(libzfs_handle_t *hdl, const char *special,
 {
 	mnttab_node_t *mtn;
 
-	(void) mutex_lock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_enter(&hdl->libzfs_mnttab_cache_lock);
 	if (avl_numnodes(&hdl->libzfs_mnttab_cache) != 0) {
 		mtn = zfs_alloc(hdl, sizeof (mnttab_node_t));
 		mtn->mtn_mt.mnt_special = zfs_strdup(hdl, special);
@@ -887,7 +888,7 @@ libzfs_mnttab_add(libzfs_handle_t *hdl, const char *special,
 		mtn->mtn_mt.mnt_mntopts = zfs_strdup(hdl, mntopts);
 		avl_add(&hdl->libzfs_mnttab_cache, mtn);
 	}
-	(void) mutex_unlock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_exit(&hdl->libzfs_mnttab_cache_lock);
 }
 
 void
@@ -896,7 +897,7 @@ libzfs_mnttab_remove(libzfs_handle_t *hdl, const char *fsname)
 	mnttab_node_t find;
 	mnttab_node_t *ret;
 
-	(void) mutex_lock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_enter(&hdl->libzfs_mnttab_cache_lock);
 	find.mtn_mt.mnt_special = (char *)fsname;
 	if ((ret = avl_find(&hdl->libzfs_mnttab_cache, (void *)&find, NULL))
 	    != NULL) {
@@ -907,7 +908,7 @@ libzfs_mnttab_remove(libzfs_handle_t *hdl, const char *fsname)
 		free(ret->mtn_mt.mnt_mntopts);
 		free(ret);
 	}
-	(void) mutex_unlock(&hdl->libzfs_mnttab_cache_lock);
+	mutex_exit(&hdl->libzfs_mnttab_cache_lock);
 }
 
 int