--- a/usr/src/lib/libzfs/common/libzfs_mount.c
+++ b/usr/src/lib/libzfs/common/libzfs_mount.c
@@ -79,11 +79,11 @@
 #include <sys/mount.h>
 #include <sys/stat.h>
 #include <sys/statvfs.h>
-#include <sys/taskq.h>
 
 #include <libzfs.h>
 
 #include "libzfs_impl.h"
+#include "utaskq.h"
 
 #include <libshare.h>
 #include <sys/systeminfo.h>
@@ -1244,7 +1244,7 @@ non_descendant_idx(zfs_handle_t **handles, size_t num_handles, int idx)
 
 typedef struct mnt_param {
 	libzfs_handle_t	*mnt_hdl;
-	taskq_t		*mnt_tq;
+	utaskq_t	*mnt_tq;
 	zfs_handle_t	**mnt_zhps; /* filesystems to mount */
 	size_t		mnt_num_handles;
 	int		mnt_idx;	/* Index of selected entry to mount */
@@ -1258,7 +1258,7 @@ typedef struct mnt_param {
  */
 static void
 zfs_dispatch_mount(libzfs_handle_t *hdl, zfs_handle_t **handles,
-    size_t num_handles, int idx, zfs_iter_f func, void *data, taskq_t *tq)
+    size_t num_handles, int idx, zfs_iter_f func, void *data, utaskq_t *tq)
 {
 	mnt_param_t *mnt_param = zfs_alloc(hdl, sizeof (mnt_param_t));
 
@@ -1270,7 +1270,7 @@ zfs_dispatch_mount(libzfs_handle_t *hdl, zfs_handle_t **handles,
 	mnt_param->mnt_func = func;
 	mnt_param->mnt_data = data;
 
-	(void) taskq_dispatch(tq, zfs_mount_task, (void*)mnt_param, TQ_SLEEP);
+	(void) utaskq_dispatch(tq, zfs_mount_task, (void*)mnt_param, TQ_SLEEP);
 }
 
 /*
@@ -1403,8 +1403,8 @@ zfs_foreach_mountpoint(libzfs_handle_t *hdl, zfs_handle_t **handles,
 	 * Issue the callback function for each dataset using a parallel
 	 * algorithm that uses a taskq to manage threads.
 	 */
-	taskq_t *tq = taskq_create("mount_taskq", mount_tq_nthr, 0,
-	    mount_tq_nthr, mount_tq_nthr, TASKQ_DYNAMIC | TASKQ_PREPOPULATE);
+	utaskq_t *tq = utaskq_create("mount_taskq", mount_tq_nthr, 0,
+	    mount_tq_nthr, mount_tq_nthr, UTASKQ_DYNAMIC | UTASKQ_PREPOPULATE);
 
 	/*
 	 * There may be multiple "top level" mountpoints outside of the pool's
@@ -1417,8 +1417,8 @@ zfs_foreach_mountpoint(libzfs_handle_t *hdl, zfs_handle_t **handles,
 		    tq);
 	}
 
-	taskq_wait(tq); /* wait for all scheduled mounts to complete */
-	taskq_destroy(tq);
+	utaskq_wait(tq); /* wait for all scheduled mounts to complete */
+	utaskq_destroy(tq);
 }
 
 /*