… usr/src/lib/libzfs/common/sys/zfs_context.h
… usr/src/lib/libzfs/common/utaskq.c
… usr/src/test/zfs-tests/tests/functional/cli_root/zfs_mount/zfs_mount_all_mountpoints.ksh