@prakashsurya 2018-02-06
@andy-js I'd appreciate your review on this, if you have the time.

I've rebased this onto the latest master changes, and the `libfakekernel` changes you made in #494. I ran into some build issues where some files would `#include <synch.h>` and this would conflict with symbols/macros from `libfakekernel`. As a workaround, I added an empty `synch.h` file in `libfakekernel`, but I'm not sure if this is the correct way to address the issue.

I still need to do some more testing, but it appears to work in the preliminary tests that I've done so far (boot, minimal zloop, mount specific tests from zfstest).