@rmustacc 2018-02-06
When trying to review this, I found the use of the fake kernel here really confusing. Why are we using it for a normal user land program? Doesn't it strike others as odd when you have a userland program use a `kmutex_t`, but then also referring to the userland bits like a normal USYNC_THREAD, which is definitely a userland mutex thing. It also makes it much less clear which mutex_enter/mutex_exit are we using. The real libc one or the renamed one in this kind of fake kernel world.

Unrelated to the above confusion, do you have an analysis somewhere about why we're marking specific file systems as requiring to be added to or removed from the mount in progress tables? It seems like this is probably an important part of this change. Can we get some comments as to why the set of file systems in question were noted as being safe to not be included in there.