@prakashsurya 2018-02-07
> Unrelated to the above confusion, do you have an analysis somewhere about why we're marking specific file systems as requiring to be added to or removed from the mount in progress tables? It seems like this is probably an important part of this change. Can we get some comments as to why the set of file systems in question were noted as being safe to not be included in there.

@rmustacc I'll open a separate PR for these changes. Looking at our internal JIRA, we have a separate issue for this with lots of details, so I think it makes sense to split them apart and land that independent of this feature. I'll work on doing that today.