@andy-js 2018-02-07
I think @rmustacc is exactly right.  We should not be building the zfs command (or libzfs) in fakekernel context.  This is a highly specialised environment that really only makes sense for things like libzpool which include kernel code.

I suggest you have a look at this earlier webrev of mine which introduces _TASKQUSER which can be used to access the taskq API outside if fakekernel context.  It was modelled after the _KMEMUSER define which does something similar:

http://cr.illumos.org/~webrev/andy_js/8115-1/