@prakashsurya 2018-02-08
@andy-js @rmustacc let me know if my latest change is what y'all were thinking. I haven't tested this yet, beyond verifying it builds.