@prakashsurya 2018-02-10
@rmustacc Thanks for the reply. Let me know after you give this version a look. If you think `libzfs` being a consumer of `libfakekernel`'s taskq implementation, then let's discuss the best way to break that out so we can do this properly.